day2: fixed implementation of part 2
The first problem was that I forgot to consume the next item from the
iterator when skipping the next entry.

The bigger problem was that in some cases, even though it's not
necessary to skip the next item in the report when viewing it from its
preceding item, it is impossible to get to future items without doing
so. An example case is the following:

`[79, 80, 83, 81, 82]`

Here, we must skip the `83`, even though we can get to it from the `80`,
because if we don't, there's no way to get to the `81` or `82` safely.

So we must always look ahead and record whether it's safe to skip the
next item, even if we don't need to do so at the moment.

Signed-off-by: Oliver Calder <oliver@calder.dev>